refactor(web): extract thumbhash canvas into Thumbhash component
